made new component for heart button
